breaking(registration): very serious refactor
For my sanity, stick with only in-memory / pmap-able version.
Also keep the coil combination by means of SoS out of the registration
business.
On the contrary, if the code is to work on ProxyArray, it needs to be
able to handle all types of the axes:
1. translation – those that are moved;
2. registration / subvolume – that which combines the subvolumes
that need to be indexed and registered
3. reduced – those, which are collapsed before any precessing
4. batch – those over which everything is looped trivially
That is too much to keep track of, especially if everything is assumed
to be completely freely parametrisable (no axis has a fixed position).

This attempts to bring sanity back to the business. It also provides
minimal abstraction for the `lmp` repo, which currently implements a
more dedicated version of the CLI, alike the one being removed in this
commit. Maybe, some part of that code will be moved back here again
later.